pyproject.toml
[project]
name = "frcnn-from-scratch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyter>=1.1.1",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "pillow>=11.3.0",
    "pyqt5>=5.15.11",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
]
